+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-04-09 16:39:33
-// Commit: chore(test): update dependencies and cleanup unused imports in test
+// Updated at: 2026-04-09 16:47:53
+// Commit: perf(test): optimize memory consumption in test list viewport
 
 class testComponentState {
-  final String commitRef = '233';
-  final String lastMessage = 'chore(test): update dependencies and cleanup unused imports in test';
-  final DateTime timestamp = DateTime.parse('2026-04-09T16:39:33+05:00');
+  final String commitRef = '234';
+  final String lastMessage = 'perf(test): optimize memory consumption in test list viewport';
+  final DateTime timestamp = DateTime.parse('2026-04-09T16:47:53+05:00');
 }